template/server/requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.5
csvw==3.7.0
cymem==2.0.13
datasets==4.6.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.133.1
ffmpy==1.0.0
filelock==3.24.3
flatbuffers==25.12.19
flexcache==0.3
flexparser==0.4
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gevent==25.9.1
google-crc32c==1.8.0
gradio==6.7.0
gradio_client==2.2.0
greenlet==3.3.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
language-tags==1.2.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
miniaudio==1.61
misaki==0.9.4
mlx==0.30.6
mlx-audio==0.4.0
mlx-embeddings==0.0.5
mlx-lm==0.30.7
mlx-metal==0.30.6
mlx-vlm==0.3.12
mlx-whisper==0.4.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.3
num2words==0.5.14
numba==0.64.0
numpy==2.4.2
onnxruntime==1.23.2
openai==2.26.0
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
pandas==3.0.1
phonemizer==3.3.0
phonemizer-fork==3.3.2
pillow==12.1.1
Pint==0.25.2
-e git+https://github.com/pipecat-ai/pipecat.git@47b463ba90c4beee0a2aaec940084c8d2c52bfba#egg=pipecat_ai
platformdirs==4.9.4
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
pylibsrtp==1.0.0
pyloudnorm==0.2.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
resampy==0.4.3
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
segments==2.4.0
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.6
uvicorn==0.41.0
vllm-mlx @ git+https://github.com/waybarrios/vllm-mlx.git@6d556311280254ce9fe7568479176e354dbe5bfe
wasabi==1.1.3
weasel==0.4.3
websocket==0.2.1
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
zope.event==6.1
zope.interface==8.2
